Refactor the second project version
